Create more ServerAlerts. Improvements to TaskParams and TaskResult
